Migration to alter spell number size

+++ plugins/admissions/models.py
@@ -211,7 +211,7 @@ class TransferHistory(models.Model):
     encounter_id                    = models.IntegerField(blank=True, null=True)
     # What is this?
     encounter_slice_id              = models.IntegerField(blank=True, null=True)
-    spell_number                    = models.IntegerField(blank=True, null=True)
+    spell_number                    = models.BigIntegerField(blank=True, null=True)
     mrn                             = models.CharField(blank=True, null=True, max_length=255)
     transfer_sequence_number        = models.IntegerField(blank=True, null=True)
     active_transfer_sequence_number = models.IntegerField(blank=True, null=True)
